Add a build script to deal with `chtype`
The size of `chtype` depends on the platform, on the ncurses' ABI
version and, in the end, on the compile-time options selected during the
compilation of the library (`--disable-lp64`, `--witch-chtype`, …).

The easiest solution is to compile a small C program to check the size
of `chtype` on the installed ncurses, hence the build script introduced
in this commit.

(This is my first use of build.rs, the code may be ugly :/ but it seems
to do the job, improvements are welcome)

For the ones interested, here is a bit of history:
Initially, `chtype` was a typedef to `long` which is, on LP64 data
model, 32-bit wide on 32-bit systems and 64-bit wide on 64-bit systems.

On 2005-03-26, ncurses developpers realized that it was wasteful to use
64-bit for `chtype` and thus changed the typedef to `unsigned` (which is
32-bit on both 32-bit and 64-bit systems).
Now, this kind of change breaks the ABI, thus they added a compile-time
option (`--disable-lp64`) to enable the new behavior.
Consequently, ncurses ABI v5 uses 64-bit `chtype` on 64-bit systems
(unless it is compiled with `--disable-lp64`) and ncurses ABI v6 use the
new behavior (use 32-bit `chtype` even on 64-bit systems) as default.

In the end, the size fo `chtype` depends on your version of ncurses and
how it was compiled (there is also an option `--with-chtype`, introduced
to support systems that doesn't use the LP64 data model).

Source: ncurses changelog