Use pkg-config for all linked libraries.
Fixes crashes when mixing wide and non-wide linker dependencies.

For example if both the "wide" and the "panel" features are enabled,
libncursesw.so and libpanel.so were linked which can not work together.

Another example is on Gentoo Linux where ncurses is split into
libncurses.so and libtinfo.so, previously libtinfo.so and libncursesw.so
were linked together if the "wide" feature was set.

Also fix the recognition of wide_chtype for non-default include paths.

Fixes #151.