fix the try_link() logic in build.rs
* on OpenBSD libmenu and libpanel both require linking with libncurses
lib, else they'll have undefined symbols.
This works on other distros/OSes because libmenu for example has a
dynamic link dependency on libncurses marked inside itself as NEEDED,
but not on OpenBSD 7.5. So try_link() will retry a failed linkage with
-lncurses too, before returning the result.
This required the order of the libs that we try(and tell cargo about to
link) to change from menu,panel,tinfo,ncurses to: ncurses first.

* also show current dir and any set env. vars for commands executed
  during build.rs (they're tipically unset/empty though, ie. inherited
  from parent)
* some fixmes added